Extract code into active_record/comments folder
